backend/e2b-requirements.txt
# =============================================================================
# E2B Sandbox Requirements
# =============================================================================
# This file contains all dependencies needed to run the tool_server inside
# the E2B sandbox. It is separate from the main requirements.txt to allow
# for a cleaner dependency resolution specific to the sandbox environment.
#
# Key difference: google-cloud-storage is pinned to <3.0 for vertexai compat
# =============================================================================

# --- Core Framework ---
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.12.0
starlette>=0.50.0

# --- HTTP Clients ---
httpx>=0.24.0
aiohttp>=3.8.0
requests>=2.31.0

# --- MCP (Model Context Protocol) ---
fastmcp>=2.14.1
mcp>=1.25.0

# --- Browser Automation ---
playwright>=1.57.0

# --- AI/ML ---
openai>=2.14.0
langchain-openai>=1.1.6
langchain-text-splitters>=1.1.0

# --- Google Cloud (CRITICAL: vertexai requires google-cloud-storage<3.0) ---
protobuf==5.29.5
grpcio==1.67.1
grpcio-status==1.67.1
proto-plus==1.26.1
google-api-core==2.27.0
googleapis-common-protos==1.66.0
google-cloud-storage==2.18.2  # Pinned to <3.0 for vertexai compatibility
gcloud-aio-storage>=9.6.1  # Async storage client (no conflict)
vertexai>=1.71.1
google-genai>=1.56.0
google-crc32c>=1.7.1
backoff>=2.0.1

# --- Web Scraping ---
beautifulsoup4>=4.12.0
lxml>=6.0.2

# --- DuckDuckGo Search ---
duckduckgo-search>=8.1.1
ddgs>=9.10.0

# --- Image Processing ---
pillow>=12.0.0

# --- Document Processing ---
pymupdf>=1.26.7
python-pptx>=1.0.2
python-docx>=1.2.0

# --- File Type Detection ---
python-magic>=0.4.27

# --- Terminal Management ---
libtmux>=0.28.0

# --- Environment ---
python-dotenv>=1.0.0

# --- Type Extensions ---
typing-extensions>=4.15.0

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.22.0

# --- LangChain (minimal for tool_server) ---
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.1
langchain-google-genai>=2.0.6

# --- Utilities ---
tenacity>=9.1.2
json-repair>=0.7.0
jinja2>=3.1.3
markdownify>=1.1.0
readabilipy>=0.3.0
regex>=2025.11.3
numpy>=2.2.6
pandas>=2.2.3
